control pago y controles en los index

+++ controllers/GestoresController.php
@@ -38,6 +38,12 @@ class GestoresController extends Controller
      */
     public function actionIndex()
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(["site/login"]); 
+        }
+        if(Persona::findOne(['idUsuario' => $_SESSION['__id']])){
+            return $this->goHome();
+        }
         $this->layout = '/main2';
         $searchModel = new GestoresSearch();
         $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
@@ -56,6 +62,12 @@ class GestoresController extends Controller
      */
     public function actionView($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(["site/login"]); 
+        }
+        if(Persona::findOne(['idUsuario' => $_SESSION['__id']])){
+            return $this->goHome();
+        }
         $this->layout = '/main2';
         return $this->render('view', [
             'model' => $this->findModel($id),
@@ -69,6 +81,12 @@ class GestoresController extends Controller
      */
     public function actionCreate()
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(["site/login"]); 
+        }
+        if(Persona::findOne(['idUsuario' => $_SESSION['__id']])){
+            return $this->goHome();
+        }
         $this->layout = '/main2';
         $model = new Gestores();
 
@@ -111,6 +129,12 @@ class GestoresController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Yii::$app->user->isGuest) {
+            return $this->redirect(["site/login"]); 
+        }
+        if(Persona::findOne(['idUsuario' => $_SESSION['__id']])){
+            return $this->goHome();
+        }
         $this->layout = '/main2';
         $this->findModel($id)->delete();
